requirements.txt
langchain_community
langchain
streamlit_chat
streamlit
python-dotenv
boto3
python-docx
streamlit-scrollable-textbox
streamlit-feedback==0.0.3